Fix: Rename component folder
