+++ src/views/Analytics/PoolView.vue
@@ -54,14 +54,14 @@
 <script>
 import { mapState } from 'pinia';
 import { useAnalyticsStore } from '../../stores/analytics.store';
-import TableTokens from './compoments/tables/TableTokens.vue';
-import PageSelector from './compoments/PageSelector.vue';
-import IconsToken from './compoments/icons/IconsToken.vue';
-import TablePools from './compoments/tables/TablePools.vue';
-import GraphCard from './compoments/graphs/GraphCard.vue';
-import GraphVolume from './compoments/graphs/GraphVolume.vue';
-import GraphLiquidity from './compoments/graphs/GraphLiquidity.vue';
-import IconsPool from './compoments/icons/IconsPool.vue';
+import TableTokens from './components/tables/TableTokens.vue';
+import PageSelector from './components/PageSelector.vue';
+import IconsToken from './components/icons/IconsToken.vue';
+import TablePools from './components/tables/TablePools.vue';
+import GraphCard from './components/graphs/GraphCard.vue';
+import GraphVolume from './components/graphs/GraphVolume.vue';
+import GraphLiquidity from './components/graphs/GraphLiquidity.vue';
+import IconsPool from './components/icons/IconsPool.vue';
 
 export default {
   name: 'TokenView',
