+++ src/views/Analytics/components/PageSelector.vue
