+++ src/views/Analytics/components/graphs/GraphCard.vue
