+++ src/views/Analytics/components/graphs/GraphLiquidity.vue
