+++ src/views/Analytics/components/icons/IconsPool.vue
