+++ src/views/Analytics/components/icons/IconsToken.vue
