+++ src/views/Analytics/components/tables/TablePools.vue
