+++ src/views/Analytics/components/tables/TableTokens.vue
